build BQ

+++ scripts/downloadOpenAlex.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
-curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
-unzip awscliv2.zip
-sudo ./aws/install
+FILE=/human-cancer-project/awscliv2.zip
 
-aws s3 sync "s3://openalex" "openalex-snapshot" --no-sign-request --region us-east-2
+if [[ -f "$FILE" ]]; then
+    echo "$FILE exist."
+else
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+fi
+
+unzip awscliv2.zip
+sudo ./aws/install --update
+# aws s3 sync "s3://openalex" "openalex-snapshot" --no-sign-request
+aws s3 sync "s3://openalex" "openalex-snapshot" --no-sign-request --delete --region us-east-2
